requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.0
requests==2.31.0
aiosqlite==0.19.0
aiohttp==3.9.1
flask==3.0.0
flask-cors==4.0.0
schedule==1.2.0